chore: point sidebar config link to internal page
