+++ src/module.ts
@@ -1,11 +1,11 @@
 import { AppPlugin } from '@grafana/data';
 import { GlobalSettings } from './types';
 import { App } from 'components/App';
-import { ConfigPage } from 'page/ConfigPage';
+import { ConfigPageWrapper } from 'components/ConfigPageWrapper';
 
 export const plugin = new AppPlugin<GlobalSettings>().setRootPage(App).addConfigPage({
   title: 'Config',
   icon: 'cog',
-  body: ConfigPage,
+  body: ConfigPageWrapper,
   id: 'config',
 });
